requirements.txt
bertopic==0.17.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
filelock==3.20.0
fsspec==2025.12.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
llvmlite==0.46.0
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.13.0
networkx==3.6.1
nltk==3.9.2
numba==0.63.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.6.1
vaderSentiment==3.3.2
